requirements.txt
astroid==2.3.2
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.1.1
django-extensions==2.2.5
django-filter==2.2.0
django-heroku==0.3.1
django-rest-multiple-models==2.1.3
djangorestframework==3.10.3
e==1.4.5
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
psycopg2==2.8.4
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
zound==1.0
